add operations/test_cast.py, move tests there
